added masterclassList and css style
