+++ frontend/src/components/MasterclassCard/MasterclassCard.jsx
@@ -1,7 +1,7 @@
 /* eslint-disable no-nested-ternary */
 /* eslint-disable react/no-array-index-key */
 /* eslint-disable react/jsx-no-comment-textnodes */
-import { useState } from 'react';
+import { useEffect, useRef } from 'react';
 import cls from './MasterclassCard.module.scss';
 
 export const MasterclassCard = ({
@@ -10,7 +10,26 @@ export const MasterclassCard = ({
   openModal,
   closeModal,
 }) => {
-  // изменить функцию тк на бэке два метода для возврата будущих и прошлых событий
+  const popupRef = useRef(null);
+
+  // Close modal if clicked outside
+  useEffect(() => {
+    const handleClickOutside = (event) => {
+      if (popupRef.current && !popupRef.current.contains(event.target)) {
+        closeModal();
+      }
+    };
+
+    if (isModalOpen) {
+      document.addEventListener('mousedown', handleClickOutside);
+    }
+    return () => {
+      if (isModalOpen) {
+        document.removeEventListener('mousedown', handleClickOutside);
+      }
+    };
+  }, [isModalOpen, closeModal]);
+
   const isPastEvent = (eventDate) => {
     return true;
   };
@@ -50,8 +69,7 @@ export const MasterclassCard = ({
               ⋮
             </button>
             {isModalOpen && (
-              // <div className={cls.container} onClick={closeModal}>
-              <div className={cls.popup} onClick={(e) => e.stopPropagation()}>
+              <div ref={popupRef} className={cls.popup}>
                 <button type="button" className={cls.changeCourseButton}>
                   Перенести
                 </button>
@@ -59,7 +77,6 @@ export const MasterclassCard = ({
                   Отменить
                 </button>
               </div>
-              // </div>
             )}
           </div>
           <p className={cls.teacher}>{masterclass.teacher}</p>
@@ -68,68 +85,3 @@ export const MasterclassCard = ({
     </div>
   );
 };
-
-const MasterclassList = ({ masterclasses }) => {
-  const [currentPage, setCurrentPage] = useState(1);
-  const itemsPerPage = 5;
-  const [openModalIndex, setOpenModalIndex] = useState(null);
-
-  const totalPages = Math.ceil(masterclasses.length / itemsPerPage);
-  const handleNextPage = () => {
-    if (currentPage < totalPages) {
-      setCurrentPage(currentPage + 1);
-    }
-  };
-
-  const currentMasterclasses = masterclasses.slice(
-    (currentPage - 1) * itemsPerPage,
-    currentPage * itemsPerPage,
-  );
-
-  const openModal = (index) => {
-    setOpenModalIndex(index);
-  };
-
-  const closeModal = () => {
-    setOpenModalIndex(null);
-    console.log('click');
-  };
-
-  return (
-    <div>
-      {currentMasterclasses.map((masterclass, index) => (
-        <MasterclassCard
-          key={masterclass.id}
-          masterclass={masterclass}
-          isModalOpen={openModalIndex === index}
-          openModal={() => openModal(index)}
-          closeModal={closeModal}
-        />
-      ))}
-      {masterclasses.length > itemsPerPage && (
-        <div className={cls.pagination}>
-          {[...Array(totalPages)].map((_, idx) => (
-            <button
-              key={idx + 1}
-              type="button"
-              className={currentPage === idx + 1 ? cls.activePage : cls.numberPage}
-              onClick={() => setCurrentPage(idx + 1)}
-            >
-              {idx + 1}
-            </button>
-          ))}
-          <button
-            type="button"
-            className={cls.buttonFurther}
-            disabled={currentPage === totalPages}
-            onClick={handleNextPage}
-          >
-            Дальше →
-          </button>
-        </div>
-      )}
-    </div>
-  );
-};
-
-export default MasterclassList;
