+++ frontend/src/components/MasterclassCard/MasterclassCard.module.scss
@@ -121,14 +121,6 @@
     cursor: pointer;
 }
 
-// .container {
-//     width: 100%;
-//     height: 100%;
-//     position: fixed;
-//     top: 0;
-//     left: 0;
-// }
-
 .popup {
     padding: 16px 20px;
     box-sizing: border-box;
@@ -179,30 +171,4 @@
     display: flex;
     align-items: center;
     justify-content: center;
-}
-
-.pagination {
-    margin-top: 32px;
-    font-family: var(--font-primary);
-    font-size: 17px;
-    font-weight: 400;
-    line-height: 1.35;
-}
-
-.numberPage {
-    background-color: transparent;
-    margin: 0 8px;
-    cursor: pointer;
-}
-
-.activePage {
-    padding: 10px;
-    background-color: var(--color-beige-secondary);
-    border-radius: 5px;
-    cursor: pointer;
-}
-
-.buttonFurther {
-    background-color: transparent;
-    cursor: pointer;
 }
